+++ src/main/java/com/grepp/nbe562team04/model/user/UserService.java
@@ -1,7 +1,6 @@
 package com.grepp.nbe562team04.model.user;
 
-import com.grepp.nbe562team04.model.achieve.AchieveRepository;
-import com.grepp.nbe562team04.model.achieve.UsersAchieve;
+import com.grepp.nbe562team04.model.achieve.AchievementService;
 import com.grepp.nbe562team04.model.auth.code.Role;
 import com.grepp.nbe562team04.model.auth.domain.Principal;
 import com.grepp.nbe562team04.model.level.LevelRepository;
@@ -9,13 +8,11 @@ import com.grepp.nbe562team04.model.level.entity.Level;
 import com.grepp.nbe562team04.model.user.dto.UserDto;
 import com.grepp.nbe562team04.model.user.entity.User;
 
-import jakarta.persistence.EntityNotFoundException;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
@@ -33,7 +30,6 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
 
 @Service
@@ -46,8 +42,7 @@ public class UserService implements UserDetailsService{
     private final PasswordEncoder passwordEncoder;
     private final ModelMapper mapper;
     private final LevelRepository levelRepository;
-    private final UsersAchieveRepository usersAchieveRepository;
-    private final AchieveRepository achieveRepository;
+    private final AchievementService achievementService;
 
     @Transactional
     public void signup(UserDto dto, Role role){
@@ -65,8 +60,6 @@ public class UserService implements UserDetailsService{
         user.setLevel(defaultLevel);
         user.setExp(0);
         user.setCreatedAt(LocalDate.now());
-        log.info("user:{}", user);
-        log.info("저장 전 최종 비밀번호: {}", user.getPassword()); // 반드시 해시 형태여야 함
         userRepository.save(user);
     }
 
@@ -87,6 +80,7 @@ public class UserService implements UserDetailsService{
                 .orElseThrow(() -> new IllegalArgumentException("사용자 없음"));
 
         user.setComment(dto.getComment());
+        user.setNickname(dto.getNickname());
 
         if (dto.getPassword() != null && !dto.getPassword().isBlank()) {
             String hashed = passwordEncoder.encode(dto.getPassword());
@@ -103,30 +97,9 @@ public class UserService implements UserDetailsService{
         }
         userRepository.save(user);
 
-        if (isTutorialCompleted(user)) {
-            giveTutorialAchievement(user.getUserId());
+        if (achievementService.isTutorialCompleted(user)) {
+            achievementService.giveTutorialAchievement(user.getUserId());
         }
-    }
-
-    @Transactional
-    public boolean giveTutorialAchievement(Long userId) {
-        Long achieveId = 1L;
-        boolean already = usersAchieveRepository.existsByUser_UserIdAndAchievement_AchieveId(userId, achieveId);
-        if (already) return false;
-
-        UsersAchieve ua = new UsersAchieve();
-        ua.setUser(userRepository.getReferenceById(userId));
-        ua.setAchievement(achieveRepository.getReferenceById(achieveId));
-        ua.setAchievedAt(LocalDateTime.now());
-        usersAchieveRepository.save(ua);
-
-        return true;
-    }
-    public boolean isTutorialCompleted(User user) {
-        return StringUtils.hasText(user.getEmail()) &&
-                StringUtils.hasText(user.getNickname()) &&
-                StringUtils.hasText(user.getComment()) &&
-                StringUtils.hasText(user.getUserImage());
     }
 
     @Transactional
