--- src/main/resources/static/js/achievementCelebrate.js
@@ -1,20 +1,0 @@
-document.addEventListener("DOMContentLoaded", function () {
-    const params = new URLSearchParams(window.location.search);
-    if (params.get("achievement") === "true") {
-        openAchievementModal();
-        document.getElementById("confettiSound").play();
-        confetti({
-            particleCount: 150,
-            spread: 80,
-            origin: { y: 0.6 }
-        });
-    }
-});
-
-function openAchievementModal() {
-    document.getElementById("achievementModal").style.display = "block";
-}
-
-// function closeModal() {
-//     document.getElementById("achievementModal").style.display = "none";
-// }
