+++ src/main/resources/templates/dashboard/dashboard.html
@@ -161,7 +161,6 @@
           목표 기업은 최대 3개까지 생성 가능합니다.
         </p>
       </a>
-
     </div>
   </main>
 </div>
